# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Abaya Production Tracking and Inventory Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /finishing:
    get:
      operationId: listFinishingRecords
      tags: [finishing]
      summary: List finishing stage records (all stages merged)
      responses:
        "200":
          description: List of finishing records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinishingRecord"
    post:
      operationId: createFinishingRecord
      tags: [finishing]
      summary: Record a finishing stage entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFinishingRecordBody"
      responses:
        "201":
          description: Created finishing record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinishingRecord"
  /finishing/{id}:
    put:
      operationId: updateFinishingRecord
      tags: [finishing]
      summary: Update finishing record (admin only — safe fields)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFinishingRecordBody"
      responses:
        "200":
          description: Updated finishing record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinishingRecord"
  /finishing/batch-info/{batchId}:
    get:
      operationId: getFinishingBatchInfo
      tags: [finishing]
      summary: Get available quantity info for finishing a batch
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Batch finishing availability info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinishingBatchInfo"

  # FINISHED GOODS
  /finished-goods:
    get:
      operationId: listFinishedGoods
      tags: [finished-goods]
      summary: List finished goods inventory
      responses:
        "200":
          description: Finished goods inventory
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinishedGoodsEntry"
    post:
      operationId: createFinishedGoodsEntry
      tags: [finished-goods]
      summary: Add finished goods to store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFinishedGoodsBody"
      responses:
        "201":
          description: Created finished goods entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinishedGoodsEntry"
  /finished-goods/{id}:
    put:
      operationId: updateFinishedGoodsEntry
      tags: [finished-goods]
      summary: Update finished goods entry (admin only — safe fields)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFinishedGoodsBody"
      responses:
        "200":
          description: Updated finished goods entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinishedGoodsEntry"
  /finished-goods/batch-info/{batchId}:
    get:
      operationId: getFinishedGoodsBatchInfo
      tags: [finished-goods]
      summary: Get available quantity for moving a batch to finished goods store
      parameters:
        - name: batchId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Batch finished-goods availability info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinishedGoodsBatchInfo"
...
components:
  schemas:
...
    FinishingRecord:
      type: object
      properties:
        id:
          type: integer
        batchNumber:
          type: string
        cuttingBatchId:
          type: integer
        stage:
          type: string
          enum: [pressing, buttons, hanger, packing, finishing]
        productCode:
          type: string
        productName:
          type: string
        sizeName:
          type: string
        colorCode:
          type: string
        colorName:
          type: string
        inputQuantity:
          type: integer
        outputQuantity:
          type: integer
        defectiveQuantity:
          type: integer
        pendingQuantity:
          type: integer
        operator:
          type: string
        processDate:
          type: string
        remarks:
          type: string
      required: [id, stage, inputQuantity, outputQuantity]
    CreateFinishingRecordBody:
      type: object
      properties:
        cuttingBatchId:
          type: integer
        inputQuantity:
          type: integer
        outputQuantity:
          type: integer
        defectiveQuantity:
          type: integer
        operator:
          type: string
        processDate:
          type: string
        remarks:
          type: string
      required: [cuttingBatchId, inputQuantity, outputQuantity, processDate]
    UpdateFinishingRecordBody:
      type: object
      properties:
        operator:
          type: string
        processDate:
          type: string
        remarks:
          type: string
    FinishingBatchInfo:
      type: object
      properties:
        batchId:
          type: integer
        batchNumber:
          type: string
        totalReceived:
          type: integer
        totalFinishingOutput:
          type: integer
        availableForFinishing:
          type: integer
      required: [batchId, totalReceived, totalFinishingOutput, availableForFinishing]
    FinishedGoodsBatchInfo:
      type: object
      properties:
        batchId:
          type: integer
        batchNumber:
          type: string
        totalFinishingOutput:
          type: integer
        totalStoredQty:
          type: integer
        availableForStore:
          type: integer
      required: [batchId, totalFinishingOutput, totalStoredQty, availableForStore]

    # FINISHED GOODS
    FinishedGoodsEntry:
      type: object
      properties:
        id:
          type: integer
        cuttingBatchId:
          type: integer
        batchNumber:
          type: string
        productId:
          type: integer
        productCode:
          type: string
        productName:
          type: string
        sizeId:
          type: integer
        sizeName:
          type: string
        colorId:
          type: integer
        colorCode:
          type: string
        colorName:
          type: string
        quantity:
          type: integer
        entryDate:
          type: string
        remarks:
          type: string
        enteredBy:
          type: string
      required: [id, quantity, entryDate]
    CreateFinishedGoodsBody:
      type: object
      properties:
        cuttingBatchId:
          type: integer
        quantity:
          type: integer
        entryDate:
          type: string
        remarks:
          type: string
      required: [cuttingBatchId, quantity, entryDate]
    UpdateFinishedGoodsBody:
      type: object
      properties:
        entryDate:
          type: string
        remarks:
          type: string
...